✨ refactor routes and controllers, use rewrites, unifying use PascalCase, api use env
