+++ frontend/package.json
@@ -18,6 +18,7 @@
     "axios": "^1.8.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "dotenv": "^16.4.7",
     "framer-motion": "^12.6.2",
     "lucide-react": "^0.486.0",
     "motion": "^12.6.2",
